Minor frontend polishing

+++ src/viewer/utils/utils_data_upload.py
@@ -27,32 +27,34 @@ def select_project(out_dir, curr_project):
     if sel_mode is None:
         return None
 
-    if sel_mode == 'Generate Demo Data':  
-        # Copy demo dirs to user folder (TODO: make this less hardcoded)
-        demo_dir_paths = [
-            os.path.join(
-                st.session_state.paths["root"],
-                "output_folder",
-                "NiChart_sMRI_Demo1",
-            ),
-            os.path.join(
-                st.session_state.paths["root"],
-                "output_folder",
-                "NiChart_sMRI_Demo2",
-            ),
-        ]
-        demo_names = []
-        for demo in demo_dir_paths:
-            demo_name = os.path.basename(demo)
-            demo_names.append(demo_name)
-            destination_path = os.path.join(
-                st.session_state.paths["out_dir"], demo_name
-            )
-            if os.path.exists(destination_path):
-                shutil.rmtree(destination_path)
-            shutil.copytree(demo, destination_path, dirs_exist_ok=True)
-        st.success(f"NiChart demonstration projects have been added to your projects list: {', '.join(demo_names)} ")
-        return
+    if sel_mode == 'Generate Demo Data':
+        generate_button = st.button("Generate")
+        if generate_button:  
+            # Copy demo dirs to user folder (TODO: make this less hardcoded)
+            demo_dir_paths = [
+                os.path.join(
+                    st.session_state.paths["root"],
+                    "output_folder",
+                    "NiChart_sMRI_Demo1",
+                ),
+                os.path.join(
+                    st.session_state.paths["root"],
+                    "output_folder",
+                    "NiChart_sMRI_Demo2",
+                ),
+            ]
+            demo_names = []
+            for demo in demo_dir_paths:
+                demo_name = os.path.basename(demo)
+                demo_names.append(demo_name)
+                destination_path = os.path.join(
+                    st.session_state.paths["out_dir"], demo_name
+                )
+                if os.path.exists(destination_path):
+                    shutil.rmtree(destination_path)
+                shutil.copytree(demo, destination_path, dirs_exist_ok=True)
+            st.success(f"NiChart demonstration projects have been added to your projects list: {', '.join(demo_names)} ")
+            return
       
     if sel_mode == 'Create New':
         sel_project = st.text_input(
